archon-exec: set the ARCHON_CURRENTDIR variable
The variable is meant to record the current directory, $PWD in a variable
to allow archon-exec to run .apk's without specifying the absolute
directory of the apk file.

Before the change, archon-exec will hang with no warning or error when
running something like...

$ archon-exec com.corporate.robbery.apk

And now it should be able to do it.